third_party/WebKit/LayoutTests/bluetooth/idl-BluetoothDiscovery.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_throws(null, function() { new BluetoothDiscovery(); },
                'the constructor should not be callable with "new"');
  assert_throws(null, function() { BluetoothDiscovery(); },
                'the constructor should not be callable');
}, 'BluetoothDiscovery IDL test');
</script>
